NewMaintenanceWindow
Worked on New Maintenance window for a bit, needed to test and added a machine class that is used to verify that list view is working. Will continue more in a bit. Still need to create an onclick for when a machine is selected, then the maintenance request box and fields.