Reorganizing Database Initializer
I am about to just make it so there are some indexed values so the collections exist. Need to research on how to require specific fields in MongoDB (Could just let the application do it but this would also allow me to not have to just find random things to index).